rewrite logger/unit_status

+++ cppmakelib/cppmakelib/utility/time.py
@@ -1,1 +1,1 @@
-time = float
+time = int
